+++ apps/backend/src/app.module.ts
@@ -13,6 +13,7 @@ import { ProfilesModule } from './profiles/profiles.module';
 import { InstitutionsModule } from './institutions/institutions.module';
 import { PermitTypesModule } from './permit-types/permit-types.module';
 import { ApplicationsModule } from './applications/applications.module';
+import { PermitsModule } from './permits/permits.module';
 import { FilesModule } from './files/files.module';
 import { NotificationsModule } from './notifications/notifications.module';
 import { BullModule } from '@nestjs/bullmq';
@@ -45,6 +46,7 @@ import { BullModule } from '@nestjs/bullmq';
     InstitutionsModule,
     FilesModule,
     ApplicationsModule,
+    PermitsModule,
     NotificationsModule,
   ],
   providers: [